rename JellyfinLibraryFolderView to JellyfinLibraryView (+Model)
replacing the old JellyfinLibraryView
