+++ BookPlayer/Jellyfin/Library Screen/JellyfinLibraryView.swift
@@ -2,31 +2,41 @@
 //  JellyfinLibraryView.swift
 //  BookPlayer
 //
-//  Created by Lysann Tranvouez on 2024-10-26.
+//  Created by Lysann Tranvouez on 2024-10-27.
 //  Copyright © 2024 Tortuga Power. All rights reserved.
 //
 
 import SwiftUI
+import Kingfisher
 
 struct JellyfinLibraryView<Model: JellyfinLibraryViewModelProtocol>: View {
   @ObservedObject var viewModel: Model
   @StateObject private var themeViewModel = ThemeViewModel()
-  @ScaledMetric private var accessabilityScale: CGFloat = 1
+  @ScaledMetric var accessabilityScale: CGFloat = 1
   
   @State private var availableSize: CGSize = .zero
-  private let itemMinSizeBase = CGSize(width: 150, height: 150)
+  private let itemMinSizeBase = CGSize(width: 100, height: 100)
   private let itemMaxSizeBase = CGSize(width: 250, height: 250)
   private let itemSpacingBase = 20.0
   
+  var navigationTitle: String {
+    switch viewModel.data {
+    case .topLevel(libraryName: let libraryName, userID: _):
+      libraryName
+    case .folder(data: let data):
+      data.name
+    }
+  }
+
   var body: some View {
     GeometryReader { geometry in
       AdaptiveVGrid(
-        numItems: viewModel.userViews.count,
+        numItems: viewModel.items.count,
         itemMinSize: adjustSize(itemMinSizeBase, availableSize: geometry.size),
         itemMaxSize: adjustSize(itemMaxSizeBase, availableSize: geometry.size),
         itemSpacing: itemSpacingBase * accessabilityScale
       ) {
-        ForEach(viewModel.userViews, id: \.id) { userView in
+        ForEach(viewModel.items, id: \.id) { userView in
           itemView(item: userView)
             .frame(minWidth: adjustSize(itemMinSizeBase, availableSize: geometry.size).width,
                    maxWidth: CGFloat.greatestFiniteMagnitude,
@@ -37,11 +47,10 @@ struct JellyfinLibraryView<Model: JellyfinLibraryViewModelProtocol>: View {
       }
     }
     .padding()
-    .navigationTitle(viewModel.libraryName)
     .environmentObject(viewModel)
-    .onAppear { viewModel.fetchUserViews() }
-    .onDisappear { viewModel.cancelFetchUserViews() }
-    .navigationTitle(viewModel.libraryName)
+    .onAppear { viewModel.fetchInitialItems() }
+    .onDisappear { viewModel.cancelFetchItems() }
+    .navigationTitle(navigationTitle)
     .toolbar {
       ToolbarItemGroup(placement: .topBarTrailing) {
         Button(
@@ -57,14 +66,10 @@ struct JellyfinLibraryView<Model: JellyfinLibraryViewModelProtocol>: View {
   
   @ViewBuilder
   private func itemView(item: JellyfinLibraryItem) -> some View {
-    let childViewModel = viewModel.createFolderViewModelFor(item: item)
-    NavigationLink {
-      NavigationLazyView(JellyfinLibraryFolderView(viewModel: childViewModel))
-    } label: {
-      JellyfinLibraryItemView<Model.FolderViewModel>(item: item)
-        .environmentObject(childViewModel)
-    }
-    .buttonStyle(PlainButtonStyle())
+    JellyfinLibraryItemView<Model>(item: item)
+      .onAppear {
+        viewModel.fetchMoreItemsIfNeeded(currentItem: item)
+      }
   }
   
   private func adjustSize(_ size: CGSize, availableSize: CGSize) -> CGSize {
@@ -74,36 +79,88 @@ struct JellyfinLibraryView<Model: JellyfinLibraryViewModelProtocol>: View {
 }
 
 class MockJellyfinLibraryViewModel: JellyfinLibraryViewModelProtocol, ObservableObject {
-  let libraryName: String = "Mock"
-  @Published var userViews: [JellyfinLibraryItem] = []
+  let data: JellyfinLibraryLevelData
+  @Published var items: [JellyfinLibraryItem] = []
   
-  func fetchUserViews() {}
-  func cancelFetchUserViews() {}
-  
-  func createFolderViewModelFor(item: JellyfinLibraryItem) -> MockJellyfinLibraryFolderViewModel {
-    let data = JellyfinLibraryLevelData.folder(data: item)
-    return MockJellyfinLibraryFolderViewModel(data: data)
+  init(data: JellyfinLibraryLevelData) {
+    self.data = data
   }
+  
+  func createFolderViewModelFor(item: JellyfinLibraryItem) -> MockJellyfinLibraryViewModel {
+    let data = JellyfinLibraryLevelData.folder(data: item)
+    return MockJellyfinLibraryViewModel(data: data)
+  }
+  
+  func fetchInitialItems() {}
+  func fetchMoreItemsIfNeeded(currentItem: JellyfinLibraryItem) {}
+  func cancelFetchItems() {}
+  
+  func createItemImageURL(_ item: JellyfinLibraryItem, size: CGSize?) -> URL? { nil }
+  
+  func beginDownloadAudiobook(_ item: JellyfinLibraryItem) {}
   
   func handleDoneAction() {}
 }
 
-#Preview {
-  let viewModel = {
-    let viewModel = MockJellyfinLibraryViewModel()
-    viewModel.userViews = [
-      JellyfinLibraryItem(id: "0", name: "View 0", kind: .userView),
-      JellyfinLibraryItem(id: "1", name: "View 1", kind: .userView),
-      //JellyfinLibraryItem(id: "2", name: "View 2", kind: .userView),
-      //JellyfinLibraryItem(id: "3", name: "View 3", kind: .userView),
-      //JellyfinLibraryItem(id: "4", name: "View 4", kind: .userView),
-      //JellyfinLibraryItem(id: "5", name: "View 5", kind: .userView),
-      //JellyfinLibraryItem(id: "6", name: "View 6", kind: .userView),
-      //JellyfinLibraryItem(id: "7", name: "View 7", kind: .userView),
-      //JellyfinLibraryItem(id: "8", name: "View 8", kind: .userView),
-      //JellyfinLibraryItem(id: "9", name: "View 9", kind: .userView),
+#Preview("top level") {
+  let model = {
+    var model = MockJellyfinLibraryViewModel(data: .topLevel(libraryName: "Mock Library", userID: "42"))
+    model.items = [
+      JellyfinLibraryItem(id: "0.0", name: "subfolder", kind: .folder),
+      JellyfinLibraryItem(id: "0.1", name: "book", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.2", name: "another book", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.3", name: "subfolder 2", kind: .folder),
+      JellyfinLibraryItem(id: "0.4", name: "book 2 with a very very long name\nmaybe even a line break?", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.5", name: "another book 2", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.6", name: "subfolder 3", kind: .folder),
+      JellyfinLibraryItem(id: "0.7", name: "book 3", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.8", name: "another book 3", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.9", name: "subfolder 2", kind: .folder),
+      JellyfinLibraryItem(id: "0.10", name: "book 2", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.11", name: "another book 2", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.12", name: "subfolder 3", kind: .folder),
+      JellyfinLibraryItem(id: "0.13", name: "book 3", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.14", name: "another book 3", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.15", name: "subfolder 2", kind: .folder),
+      JellyfinLibraryItem(id: "0.16", name: "book 2", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.17", name: "another book 2", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.18", name: "subfolder 3", kind: .folder),
+      JellyfinLibraryItem(id: "0.19", name: "book 3", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.20", name: "another book 3", kind: .audiobook),
     ]
-    return viewModel
+    return model
   }()
-  JellyfinLibraryView(viewModel: viewModel)
+  JellyfinLibraryView(viewModel: model)
+}
+
+#Preview("folder") {
+  let model = {
+    let topLevelFolder = JellyfinLibraryItem(id: "0", name: "some folder", kind: .folder)
+    var model = MockJellyfinLibraryViewModel(data: .folder(data: topLevelFolder))
+    model.items = [
+      JellyfinLibraryItem(id: "0.0", name: "subfolder", kind: .folder),
+      JellyfinLibraryItem(id: "0.1", name: "book", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.2", name: "another book", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.3", name: "subfolder 2", kind: .folder),
+      JellyfinLibraryItem(id: "0.4", name: "book 2 with a very very long name\nmaybe even a line break?", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.5", name: "another book 2", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.6", name: "subfolder 3", kind: .folder),
+      JellyfinLibraryItem(id: "0.7", name: "book 3", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.8", name: "another book 3", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.9", name: "subfolder 2", kind: .folder),
+      JellyfinLibraryItem(id: "0.10", name: "book 2", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.11", name: "another book 2", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.12", name: "subfolder 3", kind: .folder),
+      JellyfinLibraryItem(id: "0.13", name: "book 3", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.14", name: "another book 3", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.15", name: "subfolder 2", kind: .folder),
+      JellyfinLibraryItem(id: "0.16", name: "book 2", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.17", name: "another book 2", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.18", name: "subfolder 3", kind: .folder),
+      JellyfinLibraryItem(id: "0.19", name: "book 3", kind: .audiobook),
+      JellyfinLibraryItem(id: "0.20", name: "another book 3", kind: .audiobook),
+    ]
+    return model
+  }()
+  JellyfinLibraryView(viewModel: model)
 }
